Rename python problems to avoid private import warning on test files
