+++ Python/problems/problem_0004_advantage_shuffle.py
