+++ Python/problems/problem_0007_arrange_coins.py
